+++ CHANGELOG.md
@@ -5,9 +5,13 @@
 #### What's New
 - Added ability to rewind/skip tracks by swiping back/forward
 
+#### What's Improved
+- Added support for native M4A multi-value tags based on duplicate atoms
+
 #### What's Fixed
 - Fixed app restart being required when changing intelligent sorting
 or music separator settings
+- Fixed widget/notification actions not working on Android 14
 
 ## 3.2.0